+++ src/main/java/com/livingobjects/neo4j/Neo4jErrorResult.java
@@ -2,13 +2,13 @@ package com.livingobjects.neo4j;
 
 import org.codehaus.jackson.annotate.JsonProperty;
 
-public final class Neo4jError {
+public final class Neo4jErrorResult {
 
     public final String code;
 
     public final String message;
 
-    public Neo4jError(@JsonProperty("code") String code, @JsonProperty("message") String message) {
+    public Neo4jErrorResult(@JsonProperty("code") String code, @JsonProperty("message") String message) {
         this.code = code;
         this.message = message;
     }