+++ src/main/java/com/livingobjects/neo4j/iwan/model/InvalidScopeException.java
@@ -1,0 +1,8 @@
+package com.livingobjects.neo4j.iwan.model;
+
+public final class InvalidScopeException extends RuntimeException {
+
+    public InvalidScopeException(String s) {
+        super(s);
+    }
+}